Ecolabore v4.1 first charge!
Ready to install.

Just place it in your server and follow the installation instructions.

After your editions, remember to submit only "engine/", "library/", "shared/" and "templates" folders.

We know that some functionalities are not working. Its includes:

 - Shop cart procedures
 - System auto packager
 - Financial functionalities
 - Some user profile functionalities
 - Banner module
 - Video functionalities

In this time we spect to improve code standardization and visual improvements.

We will notice when other tasks will be in focus.

Remember that you can devellop components separatedly, and we prefere that you choose this way to create your own applications. They can be integrated to the main package in the future.